+++ src/utils/authentication/oauth-client-credentials-authentication-provider.spec.ts
@@ -1,5 +1,5 @@
 import { beforeEach, describe, expect, it, jest } from '@jest/globals';
-import { fetch, Response } from '../web-client.js';
+import { request } from '../web-client.js';
 import { URLSearchParams } from 'url';
 import { OauthClientCredentialsAuthenticationProvider } from './oauth-client-credentials-authentication-provider.js';
 
@@ -13,10 +13,10 @@ describe('OauthClientCredentialsAuthenticationProvider', () => {
   const REFRESH_TOKEN = 'refresh-token';
 
   let provider: OauthClientCredentialsAuthenticationProvider<TokenResponseModel>;
-  let mockFetch: jest.Mock<typeof fetch>;
+  let mockRequest: jest.Mock<typeof request>;
 
   beforeEach(() => {
-    mockFetch = fetch as jest.Mock<typeof fetch>;
+    mockRequest = request as jest.Mock<typeof request>;
 
     provider = new OauthClientCredentialsAuthenticationProvider(
       CLIENT_ID,
@@ -46,8 +46,8 @@ describe('OauthClientCredentialsAuthenticationProvider', () => {
         expires_in: expect.any(Number),
       });
 
-      expect(mockFetch).toHaveBeenCalledTimes(1);
-      expect(mockFetch).toHaveBeenCalledWith(
+      expect(mockRequest).toHaveBeenCalledTimes(1);
+      expect(mockRequest).toHaveBeenCalledWith(
         REQUEST_TOKEN_URL,
         {
           method: 'POST',
@@ -74,7 +74,7 @@ describe('OauthClientCredentialsAuthenticationProvider', () => {
         Authorization: `Bearer ${ACCESS_TOKEN}`,
       });
 
-      expect(mockFetch).toHaveBeenCalledTimes(1);
+      expect(mockRequest).toHaveBeenCalledTimes(1);
     });
 
     it('should refresh expired access token', async () => {
@@ -86,12 +86,12 @@ describe('OauthClientCredentialsAuthenticationProvider', () => {
         Authorization: `Bearer ${ACCESS_TOKEN}`,
       });
 
-      expect(mockFetch).toHaveBeenCalledTimes(2);
+      expect(mockRequest).toHaveBeenCalledTimes(2);
     });
   });
 
   function mockSuccessRequest(expiresIn: number = 3600) {
-    mockFetch.mockResolvedValueOnce({
+    mockRequest.mockResolvedValueOnce({
       ok: true,
       status: 200,
       text: () =>